pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyautogui2"
version = "1.0.0"
description = "Modern cross-platform GUI automation library - Complete rewrite of PyAutoGUI"
readme = "README.md"
requires-python = ">=3.8"
license = "BSD-3-Clause"
authors = [
    {name = "Damien A", email = "yyyymmdd.contributor@free.fr"},
]
maintainers = [
    {name = "Damien A", email = "yyyymmdd.contributor@free.fr"},
]
keywords = [
    "gui",
    "automation",
    "testing",
    "mouse",
    "keyboard",
    "cursor",
    "click",
    "press",
    "screenshot",
    "wayland",
    "x11",
    "linux",
    "windows",
    "macos",
    "cross-platform"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Desktop Environment",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "pyscreeze>=0.1.30",
    "pillow>=11.3.0",
    "pymsgbox>=1.0.9",
    "pytweening>=1.2.0",
    "pygetwindow>=0.0.5",
    "python-xlib>=0.33; sys_platform=='linux'",
    "pyobjc-core>=2.5.1; sys_platform=='darwin'",
    "pyobjc-framework-Quartz>=2.5.1; sys_platform=='darwin'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
    "pytest-retry>=1.7",
    "black>=23.0",
    "isort>=5.12",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "build>=1.0",
    "twine>=4.0",
    "graphviz>=0.21",
]

[project.urls]
Homepage = "https://github.com/D4m13n-contrib/pyautogui2"
Documentation = "https://github.com/D4m13n-contrib/pyautogui2#readme"
Repository = "https://github.com/D4m13n-contrib/pyautogui2"
"Bug Tracker" = "https://github.com/D4m13n-contrib/pyautogui2/issues"
Changelog = "https://github.com/D4m13n-contrib/pyautogui2/blob/main/CHANGELOG.md"

[project.scripts]
pyautogui2-verify = "pyautogui2.utils.cli.verify:main"
pyautogui2-install-wayland-gnomeshell-extension = "pyautogui2.osal.linux.display_servers.wayland.gnome_shell.extension.installer:main"
pyautogui2-detect-platform = "pyautogui2.osal.platform_info:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pyautogui2 = [
    "osal/linux/display_servers/wayland/gnome_shell/extension/**/*",
]

# === PYTEST ===
[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "real: marks tests requiring real OS interaction (GUI environment needed)",
    "linux: Tests specific to Linux",
    "linux_gnome_shell: Tests specific to GNOME Shell",
    "linux_wayland: Tests specific to Wayland",
    "linux_x11: Tests specific to X11",
    "macos: Tests specific to MacOS",
    "windows: Tests specific to Windows",
]

# === COVERAGE ===
[tool.coverage.run]
source = ["src/pyautogui2"]
relative_files = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/dist-packages/*",
    "*/.venv/*",
    "*/venv/*",
    "*/gnome_shell/_backend.py",    # GnomeShell backend cannot be covered on "not real" tests
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# === BLACK ===
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# === ISORT ===
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# === RUFF ===
[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "docs",
    "*.egg-info",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "D",   # docstring
    "F",   # pyflakes
    "I",   # isort (imports)
    "B",   # flake8-bugbear (bugs)
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade (modernisation)
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "S",   # bandit (security)
]
ignore = [
    "E501",   # line too long
    "S101",   # assert used
    "S603",   # subprocess without shell=True check
    "S607",   # partial executable path
    "B008",   # function call in default argument
    "D102",   # public methods (inherited docstrings from abstract methods)
    "D105",   # magic methods
    "D107",   # __init__ (see class documentation)
    "D205",   # blank line between summary and description
    "D200",   # one-line docstring
    "UP045",  # use "X | None" instead of "Optional[X]"
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",  # assert allowed in tests
    "D100",  # Module docstring
    "D101",  # Class docstring  
    "D102",  # Method docstring
    "D103",  # Function docstring
    "D104",  # Package docstring
    "ARG001",  # Unused function argument
    "ARG002",  # Unused method argument
]
"**/installer.py" = ["S603", "S607"]                # sys.exit allowed for installer scripts
"src/pyautogui2/osal/linux/__init__.py" = ["I001"]  # manual import order for clarity of logical grouping

[tool.ruff.lint.isort]
known-first-party = ["pyautogui2", "tests"]
split-on-trailing-comma = false
lines-after-imports = 2
lines-between-types = 1

[tool.ruff.lint.mccabe]
max-complexity = 10

# === MYPY ===
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

ignore_missing_imports = true
namespace_packages = true

warn_redundant_casts = true
warn_unused_ignores = false
warn_unreachable = true
strict_equality = true

show_error_codes = true
pretty = true

exclude = [
    "^build/",
    "^dist/",
    "^docs/",
    "^venv/",
    "^\\.venv/",
    "^tests/",
]

[[tool.mypy.overrides]]
module = "pyautogui2._version"
ignore_errors = true

[[tool.mypy.overrides]]
module = "pyautogui2.osal.windows.*"
disable_error_code = ["attr-defined"]

